Modified ClientConnection.retrieve() to avoid blocking when server doesn't send newline - Improved robustness with timeout handling - All client tests now get actual output instead of empty strings - Prep for reliable echo server communication
